split up config + update one-page example
